Create 'setSignatureZone()' as a module import.
